+++ src/core_services/apple_events/mod.rs
@@ -1,5 +1,7 @@
 mod desc;
 mod event_class;
+mod event_id;
 
 pub use desc::*;
 pub use event_class::*;
+pub use event_id::*;